added BaseActivity.kt
