+++ app/src/main/java/com/pocket/allinone/ui/QrCodeActivity.kt
@@ -1,27 +1,18 @@
 package com.pocket.allinone.ui
 
 import android.os.Bundle
-
-import androidx.appcompat.app.AppCompatActivity
 import com.pocket.allinone.R
-import kotlinx.android.synthetic.main.layout_toolbar.*
 
-class QrCodeActivity : AppCompatActivity() {
+class QrCodeActivity : BaseActivity() {
+    override fun getLayoutResource(): Int {
+        return R.layout.activity_qr_code
+    }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_qr_code)
-        iniToolbar()
-    }
-
-    private fun iniToolbar(){
-        setSupportActionBar(toolbar)
-        supportActionBar?.setDisplayHomeAsUpEnabled(true)
-        supportActionBar?.title=resources.getString(R.string.qr_and_barcode_scanner)
+        setTitle(resources.getString(R.string.qr_and_barcode_scanner))
 
     }
 
-    override fun onSupportNavigateUp(): Boolean {
-        onBackPressed();
-        return true;
-    }
+
 }
