+++ app/src/main/res/layout/activity_unknown_number.xml
@@ -6,18 +6,13 @@
     android:layout_height="match_parent"
     tools:context=".ui.UnknownNumberActivity">
 
-    <include
-        android:id="@+id/toolbarLayout"
-        layout="@layout/layout_toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content" />
 
     <androidx.core.widget.NestedScrollView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:fillViewport="true"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/toolbarLayout">
+        app:layout_constraintTop_toTopOf="parent">
 
         <LinearLayout
             android:layout_width="match_parent"
